requirements.txt
asgiref==3.4.0
boto3==1.17.102
botocore==1.20.102
Django==3.2.4
django-filter==2.4.0
django-storages==1.11.1
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
secure-smtplib==0.1.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
